Fixed compiler warnings

+++ src/utils.h
@@ -10,7 +10,7 @@ namespace beemaster {
     inline bool contains_string(std::string stack, std::string needle, char delimiter) {
         return stack == needle
                || stack.find(needle + delimiter) == 0 //starts with
-               || stack.find(delimiter + needle) == std::max(0, (int)(stack.length() - needle.length() -1)) //ends with
+               || stack.find(delimiter + needle) == std::max((size_t)0, (stack.length() - needle.length() - 1)) //ends with
                || stack.find(delimiter + needle + delimiter) != std::string::npos;
     }
